requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.4
django-heroku==0.3.1
django-rest-authtoken==2.1.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Pillow==9.1.1
psycopg2==2.9.3
PyJWT==2.4.0
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.1.0
